[new plugin] added lock and [new plugin] init

+++ Stella/helper/disable.py
@@ -1,5 +1,8 @@
 from Stella.helper.custom_filter import DISABLE_COMMANDS
-from Stella.helper.chat_status import isBotCan
+from Stella.helper.chat_status import (
+    isBotCan,
+    isUserAdmin
+)
 
 from Stella.database.disable_mongo import (
     get_disabled,
@@ -12,17 +15,20 @@ def disable(func):
     async def wrapper(client, message):
         if not message.command:
             return
-        chat_id = message.chat.id
-        command = message.command[0]
-        DISABLED_LIST = get_disabled(chat_id)
-        if command in DISABLED_LIST:
-            if get_disabledel(chat_id):
-                if not await isBotCan(message, permissions='can_delete_messages'):
+        if not await isUserAdmin(message, silent=True):
+            chat_id = message.chat.id
+            command = message.command[0]
+            DISABLED_LIST = get_disabled(chat_id)
+            if command in DISABLED_LIST:
+                if get_disabledel(chat_id):
+                    if not await isBotCan(message, permissions='can_delete_messages'):
+                        return
+                    await message.delete()
                     return
-                await message.delete()
-                return
+                else:
+                    return
             else:
-                return
+                await func(client, message)
         else:
             await func(client, message)
